+++ CHANGELOG.md
@@ -23,6 +23,7 @@ $ npm install @ngxs/store@dev
 - Performance: Tree-shake `isAngularInTestMode()` [#1739](https://github.com/ngxs/store/pull/1739)
 - Performance: Storage Plugin - Tree-shake `console.*` calls and expand error messages [#1727](https://github.com/ngxs/store/pull/1727)
 - CI: Bundlesize checks should run reliably [#1812](https://github.com/ngxs/store/pull/1812)
+- Fix Storage Plugin: Add storage fallback [#1820](https://github.com/ngxs/store/pull/1820)
 
 # 3.7.3 2021-12-02